requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
httpx>=0.25.0
mongoengine>=0.26.0
qdrant-client>=1.2.0
pydantic[email]>=2.4.2
email-validator>=2.0.0
pydantic-ai>=0.0.55
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth>=2.36.0
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-storage-blob==12.20.0